sharidng: update doc to only mention auto_shard experimental
There are been members of the community running sharding in production
and it's awesome. It's just the auto-sharding swift of that remains
experimental.

This patch removes the big sharding warning from the top of the
sharding overview page and better emphasises that it's the audo_shard
option that isn't ready for production use.

Change-Id: Id2c842cffad58fb6fd5e1d12619c46ffcb38f8a5