Make cmp_policy_info agree with the API behaviour
When we try to determine the correct state from a set of inconsistent
replicas of the state, the outcome should always be the same as if the
originating requests had all been handled by a single API endpoint.

In the case of storage policy index, the least recently PUT index wins
at the container server API, but previously cmp_policy_info favored
the most recently PUT index if after a DELETE.

This patch modifies cmp_policy_info to be consistent with the API
behavior.

Co-Authored-By: Clay Gerrard <clay.gerrard@gmail.com>
Change-Id: Ia107c1c8ac051890bea00770f06c53a47eafb9d6