Refactor memcache config and MemcacheRing loading
The loading and creation of the Memcache ring in the middleware is
rather interesting. It not only reads the config file, but also may
look for a `/etc/swift/memcache.conf`. Further, we are know are
looking at using the MemcacheRing client in more places.

So this patch moves the config reading from the middleware and
into a `load_memcache` helper method in swift/common/memcached.py.

Drive-by: cleanup unused stuff in middleware test module

Change-Id: I028722facfbe3ff8092b6bdcc931887a169cc49a