De-clutter container sync sample internal client conf
Since the Related-Change, the inline default internal client config
for container-sync doesn't need to be a carbon-copy of the actual
internal-client.conf-sample, so all the comment lines are removed
because this is not the place to be maintaining documentation of the
internal client config.

Also, add clarifying comment to utils.get_logger().

Related-Change: I844381fb9e1f3462043d27eb93e3fa188b206d05
Change-Id: I623d13a2d4528f28dcfb6e64407372c504206074