CI: Run ceph and rolling upgrade tests under py3
As part of that, the ceph test runner needed up-rev'ing to run under
py3. As a result, the known-failures shifted.

Trim the on-demand rolling upgrade jobs list -- now that it's running
py3, we only expect it to pass for train and beyond.

Also, pin smmap version on py2 -- otherwise, the remaining experimental
jobs running on centos-7 fail.

Change-Id: Ibe46aecf0f4461be59eb206bfe9063cc1bfff706