pytest: explicitly set system logger to DEBUG
We had a test grab the system logger via logging.getLogger() but when
running under pytest it isn't returned at the DEBUG level as it does in
nosetests.

This patch updates the test to set the level to DEBUG explicitly and
allows the only unit test that fails `pytest test/unit` to pass.

Change-Id: I1c93136cd13e927a2deb380a95fb9f96ec79fa30