Fix s3api cross policy copy
In s3api's request object we copy backend headers into the request
environ for logging after we call get_response.  The problem with s3api
copy is that we make a pre-flight HEAD request to the source object
using the same request object, so the first response backend headers
pollute the request and the proxy won't over-ride the backend header
with the correct storage policy.

As a possible fix we simply remove the problematic header from the
request object after the pre-flight HEAD request finishes.

Change-Id: I40b252446b3a1294a5ca8b531f224ce9c16f9aba