Sharding: Sync container_stat table with fresh db
When moving an UNSHARDED broker into the SHARDING db state the important
information from the retiring DB needs to be moved into the fresh DB.
However, currently we are missing a whole bunch of container information
from the retiring db.
Currently we dont sync across the delete_timestamp, status or
status_changed_at metadata either. These are important bits of information
from the retiring DB.

Delete_timestamp for instance, might indicate the UNSHARDED container is
actually deleted, but there are objects inside that need to be dealt
with. Currently, when transitioning to the SHARDING state it effectively
undeletes itself.

This patch, when moving into the SHARDING DB state via
broker.set_sharding_state() now does an SQL update to
update bits of the container_stat metadata that doesn't have direct API access
from the broker. Namely delete_timestamp, status and status_changed_at.

To get the container status, I had to add status to the get_info SQL.
The fact that it had status_changed_at and not status feels like a bug,
so adding seems to be in scope for this change.

Change-Id: Ic3b3f21ce67c640dfb632b79fd09902f14576eaf