sharder: refactor _audit_shard_container
This patch is purely a refactor. The only intended behavioral change
is the following consequence of the refactor: when the audit fails to
fetch a copy of the shard's own_shard_range from root, the warning for
that is now logged before any other shard audit warnings rather than
after.

Co-Authored-By: Clay Gerrard <clay.gerrard@gmail.com>
Change-Id: Ibc7d298fa8dfbc5c771db2e861cacc2e8af104cc