Temporary pin the swift-dsvm-functional nodeset to Focal
As per the community wide goal to migrate the CI/CD from
Ubuntu Focal to Ubuntu jammy, we need to merge the devstack, tox base
jobs to jammy on Nov 18. But swift-dsvm-functional job is
failing on Ubuntu Jammy.

To move ahead to merge the base job patches we need to pin the
swift-dsvm-functional job nodeset to Focal until this is fixed
for Jammy.

Needed-By: https://review.opendev.org/c/openstack/devstack/+/860795

Related-bug: #1996627
Change-Id: Ia35e55782592175a6a9aaafd59d72a0ff8362d61