proxy: add test for ContainerController._GET_using_cache
Add coverage for container the GET path that attempts to use memcache.

When memcache is not available the backend requests will reverse shard
ranges if required by the request params. When memcache is available
the proxy sends a 'X-Backend-Override-Shard-Name-Filter' header to
instruct the backend to always return ShardRanges in their natural
order. Add test coverage for the latter case.

Also add a last_modified item to the shard range dicts in mock
response bodies to be consistent with the real server response.

Change-Id: Ic0454c5f1d37a84258e43427cfe7cd6dfced285b