trivial: fix flakey account/test_auditor.py assertion
The test assertion assumes the test environment has exactly one
policy, but the test setup uses whatever number of policies are
currently in POLICIES. This can cause the test to fail in environments
where more than one policy is configured.

Fix the assertion to also use the actual number of policies.

Change-Id: I8fecc583f211913144663c85ab27ad448b5600f8