CI: pin tox at the project level
New tox has a bunch of incompatibilities -- on master it's worth fixing
things so we can work in the new world, but on stable branches it seems
simplest to keep ourselves in the old world.

Do it at the project level to ensure the pin applies to the
periodic-stable jobs.

Also simplify our warning suppressions. The message filter is a regex,
so any prefix of the message will suffice. This allows us to also drop a
new message seen on CentOS 8:

   CryptographyDeprecationWarning: Python 3.6 is no longer
   supported by the Python core team. Therefore, support for
   it is deprecated in cryptography. The next release of
   cryptography (40.0) will be the last to support Python 3.6.

As we've previously seen with cryptography warnings, this can slow down
our probe tests to the point that they time out.

Also, skip installing python-dev on jammy.

Change-Id: I316170442c67c1b4a5b87f9a1168cc04ca2417b8
(cherry picked from commit eb994ea501d96fdf73d4df9479a3b7d51e2d5744)