[stable-only] Pin tox<4 for stable branches (<=stable/zed) testing
tox4 need changes in tox.ini file to be compatible with it.
Those are ok to fix in master brnanch but on stable branches
testing we do not need to fix them and should keep running the
tox which was used in when those stable was released.

- https://lists.openstack.org/pipermail/openstack-discuss/2023-January/031668.html

It is pinned same way in devstack side too
- https://review.opendev.org/q/I9a138af94dedc0d8ce5a0d519d75779415d3c30b

This only pin for jobs running for stable branches <=stable/zed.

Conflicts:
 this file is not present in stable/yoga and less
 -    tools/playbooks/s3api-tests/run.yaml

Change-Id: I8b4527ffff557ab227910816a729579c06557911
(cherry picked from commit a2d363de61718d4b5a15e2c553946b71cd2464ed)