Drop tempest and grenade testing
Recently, these jobs started failing while installing tempest. The
trouble seems to be:

* tempest 26.1.0 depends on paramiko>=2.7.0
* paramiko 3.0.0 is selected which in turn depends on
  cryptography >= 3.3 and bcrypt >= 3.2
* cryptography 39.0.0 and bcrypt 4.0.1 are selected, neither of which
  get installed from binary wheels and both of which require a Rust
  toolchain to compile

For some unknown reason, this used to be OK-ish -- the stable/train
upper-constraints file would be respected, paramiko==2.6.0 would be
installed with cryptography==2.8 and bcrypt==3.1.7, and pip would
simply warn

    ERROR: tempest 26.1.0 has requirement paramiko>=2.7.0, but you'll
    have paramiko 2.6.0 which is incompatible.

even though everything would still work. Following the paramiko 3.0.0
release, however, it stopped working and it's not entirely clear how to
fix it. Given the age of the branch and its extended-maintenance state,
dropping the testing seems best.

Change-Id: Ie22aabb0f6bde0d7f5df71de527d3201ba605181