Don't clear x-container-sysmeta-sharding on delete_db
In the container backend we have a delete_meta_whitelist which
whitelists specific metadata from being cleared when we delete a broker.
Currently the sharding root and quoted root are in there.

They're there because we now process deleted shards and we need to know
what a shard's root is. But since that time we've found and
edge case bug where we've found old sharding brokers that are stuck
because they're deleted (a deleted shard will clear the sysmeta-sharding
truth value), but also has lost their shard-ranges. This makes them fail
the sharding_enabled check and therefore never get a chance to pull new
shard ranges from their root and are therefore stuck.

A deleted shard wont be recreated, so it doesn't hurt to keep this
sharding sysmeta value, and will be an extra line of defence stopping
similar stuck container issues in the future.

Change-Id: I0bef534eca71b9ce2b29927021b1977463ffbe74