CI: fix lower-constraints job
Switch from focal back to bionic so tox gets installed with py36. This
ensures that we get virtualenv<20.18.0 so we can still create py27
environments (support for which was dropped in virtualenv==20.22.0).

Note that this also means we can drop our tox<4 pin, as newer tox does
not support py36, either.

Change-Id: Ia2b3cc4f719347705dec15f1fc97d982e72bfb34